Make the package reproducible
Whilst working on the Reproducible Builds effort [0] we noticed
that python-etcd3gw could not be built reproducibly.

This is because one of the function signatures used a non-
deterministic / random default value that was rendered by the
documentation system at build time and thus varied each time.

Patch attached.

 [0] https://reproducible-builds.org/

Note: The original author of this patch is Chris Lamb <lamby@debian.org>